to-be-expert-main/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
billiard==4.2.1
boto3==1.36.26
botocore==1.36.26
Brotli==1.1.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.1
cssselect2==0.7.0
defusedxml==0.7.1
distro==1.9.0
dj-database-url==2.3.0
Django==5.1.2
django-cors-headers==4.5.0
django-debug-toolbar==4.4.6
django-environ==0.12.0
django-filter==24.3
django-oauth-toolkit==3.0.1
django-storages==1.14.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-social-oauth2==3.1.0
drf-yasg==1.21.7
fonttools==4.55.3
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
idna==3.10
inflection==0.5.1
jiter==0.6.1
jmespath==1.0.1
jwcrypto==1.5.6
kombu==5.4.2
minio==7.2.12
oauthlib==3.2.2
openai==1.51.2
packaging==24.1
pillow==10.4.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pydyf==0.11.0
PyJWT==2.9.0
pyparsing==3.2.1
pyphen==0.17.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.11.2
sendgrid==6.11.0
setuptools==75.1.0
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
starkbank-ecdsa==2.2.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
weasyprint==63.1
webencodings==0.5.1
whitenoise==6.8.2
zopfli==0.2.3.post1
